refactor: add detail-panel-header component for and extract components
